+++ mobile/App.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Provider } from 'react-redux'
 
-import NavigationWithRedux from './src/navigation'
+import NavigationWithRedux from './src/views'
 import redux from './src/redux'
 
 const App = () => (
